Work around signtool.exe's timestamping stupidity
One might reason that "/tp 0" adds a timestamp to the first signature
in a file. Instead signtool.exe rejects 0 with "invalid index". This
is odd, as "/tp 1" works for the _second_ signature. This commit
works around this feature by not defining "/tp" at all, except when
adding timestamps to secondary signatures.

Signed-off-by: Samuli Seppänen <samuli@openvpn.net>